refactor(day-planner): move model classes #32
