+++ app/src/main/java/pl/gawor/android/tayckner/common/model/CredentialsModel.kt
@@ -1,3 +1,3 @@
-package pl.gawor.android.tayckner.habit_tracker.model
+package pl.gawor.android.tayckner.common.model
 
 data class CredentialsModel(val username: String, val password: String)
