+++ app/src/main/java/pl/gawor/android/tayckner/habit_tracker/model/Habit.kt
@@ -1,5 +1,7 @@
 package pl.gawor.android.tayckner.habit_tracker.model
 
+import pl.gawor.android.tayckner.common.model.User
+
 data class Habit(
     var id: Long,
     var name: String,
